+++ lib/gpui/native/generated.ex
@@ -22,6 +22,10 @@ defmodule GPUI.Native.Generated do
         :erlang.nif_error(:nif_not_loaded)
       end
 
+      def set_theme(_runtime, _mode) do
+        :erlang.nif_error(:nif_not_loaded)
+      end
+
       def put_resource(_runtime, _resource_id, _resource) do
         :erlang.nif_error(:nif_not_loaded)
       end